Remove the original notebook
I have renamed it to match the specific problem.  Since jupyter
notebooks are insanely verbose xml, there is not much point
trying to track the file.  This is not really code, and should
not be in a git repo.  (Neither should the data file, but this
seems to be the way the modern world is trending; throw all your
crap into a git repo and let god sort it out.  sigh.)