Remove the colab generated python
Although it is easier to track the history of changes using the python
file, it is anathema to include any auto-generated code in a git repo.
Since we are using this git repo as a public presentation layer, we need
to make some compromises and will keep the Jupyter notebook in the repo
(with reservations).